Move opal/replutils to opal-replutils
