+++ stdlib/opal-replutils.rb
