+++ stdlib/opal/repl_js.rb
@@ -1,5 +1,5 @@
 require 'opal/platform'
-require 'opal/replutils'
+require 'opal-replutils'
 require 'json'
 
 REPLUtils.js_repl
